move all `Regex` from `csontants` into separate module
