+++ src/api_types/src/oidc.rs
@@ -3,7 +3,7 @@ use crate::generic::PasswordPolicyResponse;
 use crate::sessions::SessionState;
 use actix_web::HttpRequest;
 use actix_web::http::header;
-use rauthy_common::constants::{
+use rauthy_common::regex::{
     RE_ALNUM, RE_BASE64, RE_CLIENT_ID_EPHEMERAL, RE_CODE_CHALLENGE_METHOD, RE_CODE_VERIFIER,
     RE_GRANT_TYPES, RE_LOWERCASE, RE_SCOPE_SPACE, RE_URI,
 };
