+++ src/api_types/src/scopes.rs
@@ -1,5 +1,5 @@
 use crate::cust_validation::validate_vec_attr;
-use rauthy_common::constants::RE_GROUPS;
+use rauthy_common::regex::RE_GROUPS;
 use serde::{Deserialize, Serialize};
 use utoipa::ToSchema;
 use validator::Validate;
